+++ src/structs/mod.rs
@@ -1,6 +1,6 @@
 //! Data structures and implementations.
 
-pub use self::dmat::DMat;
+pub use self::dmat::{DMat, DMat1, DMat2, DMat3, DMat4, DMat5, DMat6};
 pub use self::dvec::{DVec, DVec1, DVec2, DVec3, DVec4, DVec5, DVec6};
 pub use self::vec::{Vec0, Vec1, Vec2, Vec3, Vec4, Vec5, Vec6};
 pub use self::pnt::{Pnt0, Pnt1, Pnt2, Pnt3, Pnt4, Pnt5, Pnt6};
@@ -11,6 +11,7 @@ pub use self::persp::{Persp3, PerspMat3};
 pub use self::ortho::{Ortho3, OrthoMat3};
 pub use self::quat::{Quat, UnitQuat};
 
+mod dmat_macros;
 mod dmat;
 mod dvec_macros;
 mod dvec;
